requirements.txt
Django>=4.2
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
Pillow>=10.0          # для ImageField
mysqlclient>=2.2      # MySQL драйвер (або PyMySQL як альтернатива)
